+++ bugsnag-android-core/src/test/java/com/bugsnag/android/StacktraceSerializationTest.kt
@@ -22,7 +22,7 @@ internal class StacktraceSerializationTest {
                 Stacktrace(arrayOf(), emptySet(), NoopLogger),
 
                 // empty custom frames ctor
-                Stacktrace(listOf(frame)),
+                Stacktrace(mutableListOf(frame)),
 
                 // basic
                 basic(),
@@ -57,7 +57,7 @@ internal class StacktraceSerializationTest {
         }
 
         private fun trimStacktraceListCtor(): Stacktrace {
-            val elements = (0..999).map { count ->
+            val elements = (0..999).mapTo(ArrayList()) { count ->
                 Stackframe("Foo", "Bar.kt", count, true).also { frame ->
                     // set different type for each frame
                     frame.type = when (count % 3) {